requirements.txt
openai
assemblyai
moviepy
python-dotenv
